+++ src/lib.rs
@@ -26,8 +26,8 @@
 //! let ray = Ray::new(origin, direction);
 //!
 //! struct Sphere {
-//!     position: Point3<f32>,
-//!     radius: f32,
+//!     position: Point3<f64>,
+//!     radius: f64,
 //!     node_index: usize,
 //! }
 //!
@@ -52,8 +52,8 @@
 //!
 //! let mut spheres = Vec::new();
 //! for i in 0..1000u32 {
-//!     let position = Point3::new(i as f32, i as f32, i as f32);
-//!     let radius = (i % 10) as f32 + 1.0;
+//!     let position = Point3::new(i as f64, i as f64, i as f64);
+//!     let radius = (i % 10) as f64 + 1.0;
 //!     spheres.push(Sphere {
 //!         position: position,
 //!         radius: radius,
@@ -88,7 +88,7 @@ extern crate rand;
 
 /// A minimal floating value used as a lower bound.
 /// TODO: replace by/add ULPS/relative float comparison methods.
-pub const EPSILON: f32 = 0.00001;
+pub const EPSILON: f64 = 0.00001;
 
 pub mod aabb;
 pub mod axis;
